Ajout de méthode de calcul entre 2 villes et recup meilleure solution parmis un groupe de solution + commencement ajout de commentaire

+++ Model/Ville/Ville.cpp
@@ -30,11 +30,11 @@ int Ville::getId() {
     return id;
 }
 
-int Ville::getLatitude() {
+double Ville::getLatitude() {
     return latitude;
 }
 
-int Ville::getLongitude() {
+double Ville::getLongitude() {
     return longitude;
 }
